requirements.txt
# UrbanFlow CDMX — Dependencias de producción
# Python 3.11
# Generado: 2026-03-15

# ──────────────────────────────────────────────
# Cómputo numérico y científico
# ──────────────────────────────────────────────
numpy==1.26.4
scipy==1.13.1
pandas==2.2.2
pyarrow==16.1.0          # Backend columnar para pandas (lectura de parquet)

# ──────────────────────────────────────────────
# Datos geoespaciales
# ──────────────────────────────────────────────
geopandas==0.14.4
shapely==2.0.4
pyproj==3.6.1             # Reproyección CRS (WGS84 ↔ EPSG:6372)
fiona==1.9.6              # I/O de shapefiles
rtree==1.4.1              # Índice espacial para joins geográficos (1.2.0 no compatible con Win/Py3.12)

# ──────────────────────────────────────────────
# Machine Learning y Forecasting
# ──────────────────────────────────────────────
xgboost==2.0.3
scikit-learn==1.5.1       # Preprocesamiento, métricas, pipelines
prophet==1.1.5            # Forecasting de series de tiempo

# ──────────────────────────────────────────────
# API REST
# ──────────────────────────────────────────────
fastapi==0.111.1
uvicorn[standard]==0.30.1
pydantic==2.7.4           # Validación de schemas (incluido en FastAPI)
pydantic-settings==2.3.4  # Gestión de configuración vía .env
httpx==0.27.0             # Cliente HTTP async (interno y para APIs externas)

# ──────────────────────────────────────────────
# Dashboard
# ──────────────────────────────────────────────
streamlit==1.36.0
plotly==5.22.0            # Gráficas interactivas en Streamlit
folium==0.17.0            # Mapas interactivos
streamlit-folium==0.21.0  # Integración Folium ↔ Streamlit
altair==5.3.0             # Visualizaciones declarativas

# ──────────────────────────────────────────────
# Base de datos — PostgreSQL + PostGIS
# ──────────────────────────────────────────────
sqlalchemy==2.0.31
psycopg2-binary==2.9.9    # Driver PostgreSQL (binario para facilidad de instalación)
alembic==1.13.2           # Migraciones de esquema
geoalchemy2==0.15.2       # Tipos PostGIS para SQLAlchemy

# ──────────────────────────────────────────────
# Caché y tareas asíncronas
# ──────────────────────────────────────────────
redis==5.0.7
celery==5.4.0
celery[redis]==5.4.0      # Broker Redis para Celery
flower==2.0.1             # UI de monitoreo de Celery

# ──────────────────────────────────────────────
# Ingesta de datos y conectores externos
# ──────────────────────────────────────────────
requests==2.32.3          # Llamadas HTTP síncronas (ingesta batch)
aiohttp==3.9.5            # Llamadas HTTP async (GTFS-RT tiempo real)
gtfs-realtime-bindings==1.0.0  # Parseo de feeds GTFS-RT (Metrobús)
protobuf==4.25.3          # Dependencia de gtfs-realtime-bindings

# ──────────────────────────────────────────────
# Voz (TTS/STT para VialAI)
# ──────────────────────────────────────────────
openai>=1.40.0                # TTS (tts-1) y STT (Whisper) para interfaz de voz

# ──────────────────────────────────────────────
# Configuración y utilidades
# ──────────────────────────────────────────────
python-dotenv==1.0.1      # Carga de variables de entorno desde .env
python-multipart==0.0.9   # Soporte multipart/form-data en FastAPI
tenacity==8.4.2           # Reintentos con backoff (llamadas a APIs externas)
loguru==0.7.2             # Logging estructurado
tqdm==4.66.4              # Barras de progreso para ingestas largas
pytz==2024.1              # Manejo de zonas horarias (America/Mexico_City)

# ──────────────────────────────────────────────
# Serialización y formatos
# ──────────────────────────────────────────────
orjson==3.10.6            # JSON rápido (usado por FastAPI para responses)
openpyxl==3.1.5           # Lectura de archivos .xlsx de SEMOVI/C5
